refactor: remove webpack 4 file watcher api
