+++ lib/cached-child-compiler.js
@@ -26,7 +26,7 @@
 /** @typedef {import("webpack/lib/Compiler.js")} WebpackCompiler */
 /** @typedef {import("webpack/lib/Compilation.js")} WebpackCompilation */
 /** @typedef {{hash: string, entry: any, content: string }} ChildCompilationResultEntry */
-/** @typedef {import("./webpack4/file-watcher-api").Snapshot} Snapshot */
+/** @typedef {import("./file-watcher-api").Snapshot} Snapshot */
 /** @typedef {{fileDependencies: string[], contextDependencies: string[], missingDependencies: string[]}} FileDependencies */
 /** @typedef {{
   dependencies: FileDependencies,
